refactor(text): use size_t instead of int for cursor

+++ include/util.h
@@ -3,5 +3,5 @@
 
 #include <glib.h>
 
-void string_remove_at(char *str, int pos);
-gchar *string_insert_chars_at(gchar *str, gchar *chars, int pos);
+void string_remove_at(char *str, size_t pos);
+gchar *string_insert_chars_at(gchar *str, gchar *chars, size_t pos);
